# docs/architecture/openapi.yaml
openapi: 3.1.0
info:
  title: Maricho API
  description: Backend API for the Maricho skilled work platform.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /crews/me:
    post:
      summary: Create the current AGGREGATOR's crew
      description: One crew per lead (enforced by a unique constraint).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Crew created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crew'
        '403':
          description: Authenticated user is not an AGGREGATOR
        '409':
          description: Already has a crew registered
    get:
      summary: View the current AGGREGATOR's crew and its members
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Crew found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crew'
        '403':
          description: Authenticated user is not an AGGREGATOR
        '404':
          description: No crew registered yet
  /crews/me/members:
    post:
      summary: Add a worker to the current AGGREGATOR's crew
      description: >
        The crew lead directly manages their roster — there is no invite/
        accept sub-workflow.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [worker_id]
              properties:
                worker_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crew'
        '403':
          description: Authenticated user is not an AGGREGATOR
        '404':
          description: No crew registered yet
        '409':
          description: Worker is already a member of this crew
        '422':
          description: worker_id does not refer to a registered worker
  /crews/me/members/{worker_id}:
    delete:
      summary: Remove a worker from the current AGGREGATOR's crew
      security:
        - bearerAuth: []
      parameters:
        - name: worker_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crew'
        '403':
          description: Authenticated user is not an AGGREGATOR
        '404':
          description: No crew registered, or worker is not a member
  /crew-orders/request:
    post:
      summary: Request a crew (Crew Hire)
      description: >
        An organization requests N workers of a trade for a site. Stage Two
        scope (business case names this a self-service "Crew Hire console");
        built ahead of the phased rollout at the user's request.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trade, suburb, address, workers_needed]
              properties:
                trade:
                  type: string
                  enum: [PLUMBING, ELECTRICAL, SOLAR, WELDING, BUILDING, APPLIANCE_REPAIR]
                suburb:
                  type: string
                address:
                  type: string
                landmark_narrative:
                  type: string
                  nullable: true
                  description: >-
                    Optional navigation description, gated the same as
                    address. See ADR-004.
                latitude:
                  type: number
                  nullable: true
                  minimum: -90
                  maximum: 90
                longitude:
                  type: number
                  nullable: true
                  minimum: -180
                  maximum: 180
                workers_needed:
                  type: integer
                  minimum: 1
                problem_description:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Crew order requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrewOrder'
        '403':
          description: Authenticated user is not a BUYER
        '422':
          description: >-
            Invalid input, unknown suburb, or out-of-range
            latitude/longitude
  /crew-orders/{order_id}:
    get:
      summary: Fetch a crew order
      description: >
        Restricted to the order's buyer, the booked crew's lead, or an OPS
        user.
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Crew order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrewOrder'
        '403':
          description: Authenticated user is not a party to this crew order
        '404':
          description: Crew order not found
  /crew-orders/{order_id}/matches:
    get:
      summary: Run the crew matching engine and return ranked crew candidates
      description: >
        Returns up to 3 crews with enough qualifying members (matching
        trade, within distance, individually available) to fill the order,
        ranked by distance band then qualifying-member count. The first
        call on a REQUESTED order transitions it to MATCHED. Restricted to
        the order's buyer.
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Up to 3 ranked crew candidates (possibly empty)
          content:
            application/json:
              schema:
                type: array
                maxItems: 3
                items:
                  $ref: '#/components/schemas/CrewMatchCandidate'
        '403':
          description: Authenticated user is not this order's buyer
        '404':
          description: Crew order not found
        '409':
          description: Crew order is no longer accepting matches
  /crew-orders/{order_id}/book:
    post:
      summary: Book a matched crew
      description: >
        Holds the deposit (DEPOSIT_HELD). The crew becomes unavailable as a
        whole for other orders until this one completes. `currency`
        (ADR-002, default USD) is set once here and reused by every later
        ledger entry. When `currency` is `ECOCASH`, a real collection
        request is initiated from the buyer before booking completes.
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [crew_id, deposit_amount]
              properties:
                crew_id:
                  type: string
                  format: uuid
                deposit_amount:
                  type: number
                  exclusiveMinimum: 0
                currency:
                  type: string
                  enum: [USD, ECOCASH, ZWG]
                  default: USD
      responses:
        '200':
          description: Crew order booked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrewOrder'
        '403':
          description: Authenticated user is not this order's buyer
        '404':
          description: Crew order not found
        '409':
          description: Order is not MATCHED, or the crew is already committed elsewhere
        '422':
          description: crew_id does not refer to a registered crew
        '502':
          description: currency was ECOCASH and the collection request failed
        '503':
          description: currency was ECOCASH and no gateway config exists yet
  /crew-orders/{order_id}/quote:
    post:
      summary: Booked crew's lead submits their on-site price
      description: >
        Two-tier quoting (ADR-003): the crew lead submits a labor amount and
        a materials amount separately; quote_amount is computed server-side
        as their sum, never accepted directly from the client. Commits the
        remaining balance (BALANCE_COMMITTED) and moves the order to
        IN_PROGRESS. Only the booked crew's lead may call this.
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [labor_amount, materials_amount]
              properties:
                labor_amount:
                  type: number
                  exclusiveMinimum: 0
                materials_amount:
                  type: number
                  minimum: 0
      responses:
        '200':
          description: Quote submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrewOrder'
        '403':
          description: Authenticated user is not the booked crew's lead
        '404':
          description: Crew order not found
        '409':
          description: Order is not BOOKED
        '422':
          description: Quote does not exceed the deposit already held
  /crew-orders/{order_id}/complete:
    post:
      summary: Buyer signs off and releases payment
      description: >
        Writes a RELEASED ledger entry (the full quote amount) and sets
        `completion_note`. No per-worker RecordEntry is produced for crew
        orders. Only the order's buyer may call this.
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [what_was_done]
              properties:
                what_was_done:
                  type: string
      responses:
        '200':
          description: Crew order completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrewOrder'
        '403':
          description: Authenticated user is not this order's buyer
        '404':
          description: Crew order not found
        '409':
          description: Order is not IN_PROGRESS
  /crew-orders/{order_id}/ledger:
    get:
      summary: View a crew order's ledger entries
      description: >
        Restricted to the order's buyer, the booked crew's lead, or an OPS
        user.
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Ledger entries, oldest first, paginated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CrewOrderLedgerEntry'
        '403':
          description: Authenticated user is not a party to this crew order
        '404':
          description: Crew order not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 200
      description: Maximum rows to return.
    offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Rows to skip before collecting the result set.
  schemas:
...
    Crew:
      type: object
      properties:
        id:
          type: string
          format: uuid
        lead_id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        member_ids:
          type: array
          items:
            type: string
            format: uuid
    CrewOrder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        buyer_id:
          type: string
          format: uuid
        crew_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [REQUESTED, MATCHED, BOOKED, IN_PROGRESS, COMPLETED, CANCELLED]
        trade:
          type: string
          enum: [PLUMBING, ELECTRICAL, SOLAR, WELDING, BUILDING, APPLIANCE_REPAIR]
        suburb:
          type: string
        address:
          type: string
          description: Only present for the buyer, the booked crew's lead, or OPS.
        landmark_narrative:
          type: string
          nullable: true
          description: >-
            Optional navigation description (e.g. "blue gate opposite the
            clinic"), gated the same as address.
        latitude:
          type: number
          nullable: true
          description: Optional GPS pin, gated the same as address.
        longitude:
          type: number
          nullable: true
          description: Optional GPS pin, gated the same as address.
        workers_needed:
          type: integer
        problem_description:
          type: string
          nullable: true
        currency:
          type: string
          enum: [USD, ECOCASH, ZWG]
          nullable: true
          description: Set once at booking (ADR-002); nullable until then.
        labor_amount:
          type: number
          nullable: true
          description: >-
            Set together with materials_amount by
            POST /crew-orders/{id}/quote (ADR-003).
        materials_amount:
          type: number
          nullable: true
        quote_amount:
          type: number
          nullable: true
          description: Computed server-side as labor_amount + materials_amount.
        completion_note:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CrewMatchCandidate:
      type: object
      properties:
        crew_id:
          type: string
          format: uuid
        crew_name:
          type: string
          nullable: true
        lead_id:
          type: string
          format: uuid
        qualifying_members:
          type: integer
        total_members:
          type: integer
        distance_band:
          type: string
          enum: [SAME_SUBURB, NEARBY, MODERATE, FAR]
        distance_km:
          type: number
        reason:
          type: string
    CrewOrderLedgerEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        crew_order_id:
          type: string
          format: uuid
        amount:
          type: number
        entry_type:
          type: string
          enum: [DEPOSIT_HELD, BALANCE_COMMITTED, RELEASED, REFUNDED]
        currency:
          type: string
          enum: [USD, ECOCASH, ZWG]
        external_reference:
          type: string
          nullable: true
          description: Gateway transaction reference (ADR-002), set only for ECOCASH entries.
        created_at:
          type: string
          format: date-time
...